requirements.txt
fastapi==0.116.1
uvicorn==0.23.2
ipython==8.12.3
langchain==0.3.27
langchain_community==0.3.29
langchain_core==0.3.76
langchain_openai==0.3.33
langgraph==0.6.7
pydantic==2.11.9
python-dotenv==1.1.1
typing_extensions==4.15.0
pypdf==6.0.0

